feat: implement TIME IN/TIME OUT workflow with enhanced UX
- Add smart session detection (auto TIME IN/TIME OUT based on existing open sessions)
- Implement professional alerts (critical red for UNFIT, success green for operations)
- Add server-configurable health declaration window timeout
- Create attendance_log_with_employee composite model for richer data
- Update stored procedure to join time_logs with contractor_employee
- Enhance UI to display detailed contractor information with age calculation
- Add provider_name and project_name to contractor_employee model

BREAKING CHANGE: Requires manual execution of update_sp_time_logs_GetAllFiltered.sql

+++ Services/ISystemConfigService.cs
@@ -5,9 +5,9 @@ namespace ContractorAttendanceWithHealthDeclaration.Services
 {
     public interface ISystemConfigService
     {
-        Task<Response<int>> GetDebounceThresholdMinutes();
+        Task<Response<double>> GetDebounceThresholdMinutes();
         Task<Response<int>> GetHealthDeclarationValidityHours();
-        Task<Response<int>> GetHealthDeclarationWindowMinutes();
+        Task<Response<double>> GetHealthDeclarationWindowMinutes();
         Task<Response<string>> GetAdminADGroup();
         Task<Response<system_config>> GetByKey(string key);
         Task<Response<bool>> Update(system_config config);
